// src/app/model/roleLimit.model.ts
export class roleLimit  {  

      
itbId: number;
roleId: number;
roleName: string;
currencyIso: null;
creditLimit     : string;;
debitLimit: string;;
gLDebitLimit: string;;
gLCreditLimit: string;;
dateCreated: string;
status: string;
userId: null;
createdBy: string;
loginUserId: number;    
}



////




export class UserLimit  {  

      
        itbId: number;
        roleId: number;
        userItbId: number;
        fullName: string;
        currencyIso: null;
        creditLimit: string;;
        debitLimit: string;;
        gLDebitLimit: string;;
        gLCreditLimit: string;;
        dateCreated: string;
        status: null;
        userId: null;
        createdBy: string;
        loginUserId: number;    
}
        